Remove superfluous actions from startup.  Call update_status_line as part of Buffer#display .
This is a semantic cherry pick of 42e021f and 5c8538 .  It is not necessary
to update the context and status lines in the startup code, because the
very displaying of a buffer updates these.

switch_to_buffer_number is still necessary under some conditions (causing
nil-related problems), so it couldn't be completely removed.